pyproject.toml
[tool.poetry]
name = "eurocar-telegram-bot"
version = "0.1.0"
description = ""
authors = ["blowyourbra1n <blowyourbra1n@mail.ru>"]
readme = "README.md"
packages = [{include = "eurocar_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
load-dotenv = "^0.1.0"
aiogram = {url = "https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
